chore : Renamed the Main page to Brand page.
