+++ front/src/common2.js
@@ -1,9 +1,9 @@
 // 프로젝트에서 공통으로 사용하는 변수들을 모아둔 곳.
 
 export const AppPage = {
-  meeting: "/",
+  home: "/", // Meeting page
   meetingDetail: "/meeting/detail",
-  home: "/brand",
+  brand: "/brand",
   auth: "/auth",
   mypage: "mypage",
   search: "/search",
