+++ front/src/component/brand/MainIntroContentLeft.jsx
