+++ front/src/component/navbar/MenuList.jsx
@@ -4,9 +4,9 @@ import { AppPage } from "../../common2";
 export default function MenuList() {
   return (
     <ul className="lg:w-[50%] sm:w-[58%] hidden sm:flex justify-between text-[#344054] dark:text-white lg:text-[1.25rem] md:text-[1.16rem] sm:text-[1.1rem] text-[1.1rem]">
-      <Menu path={AppPage.meeting} title="Meeting" />
+      <Menu path={AppPage.home} title="Meeting" />
       <Menu path={AppPage.toolkit} title="Toolkit" />
-      <Menu path={AppPage.home} title="Brand" />
+      <Menu path={AppPage.brand} title="Brand" />
     </ul>
   );
 }
